Fix >512 DMX channels
DMX Universes were off by one or so which caused issues mapping >170 RGB via sACN. This changes the logic from parsing the colors while iterating over the DMX Universes, to preprocessing colors and just making the packets for the DMX universes. The issue occurred as the input color array was the number of pixels, not the individual colors. Decoding likely resulted in, when on universe 0, for LED starting at channel 511, 512, number 513 is lost and thus the error.